pyproject.toml
[project]
name = "news-sender-cron"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "schedule",
    "google-genai",
    "python-dotenv>=1.2.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
